Translated more things
I added but did not finish the enum for token types. I continued working with the leixcal anlyzer but for the file reading portion am unsure if it does what we need it to do. For the token_ID function I creted a function for it but also have not finished doing it.